+++ CHANGES.md
@@ -9,6 +9,8 @@
 
 - use pynvx as pynvml replacement on mac OS (@phenomax)
 - switched GPUMemory tuple order to be consistent with nvidia meminfo
+- handle the situation where a mem peak thread gets delayed to finish
+- some self.namespace.shell.user_ns vars disappear, so do a safe delete
 
 ## 0.1.11 (2019-01-25)