+++ CHANGES.md
@@ -5,7 +5,7 @@
 
 
 
-## 0.1.12.dev0 (Work In Progress)
+## 0.1.12 (2019-02-22)
 
 - use pynvx as pynvml replacement on mac OS (@phenomax)
 - switched GPUMemory tuple order to be consistent with nvidia meminfo